更新

+++ 普中51板/lcd1602显示复刻实验/Public/public.h
@@ -1,12 +1,12 @@
 # ifndef _public_H
 # define _public_H
 
-# include "reg52.h"
+# include <reg52.h>
 
 # define u8 unsigned char
 # define u16 unsigned int
 
 //延时函数 ms
-void delay_ms(u16 z);
+void delay_ms(u8 z);
 
 # endif
